requirements.txt
requests~=2.26.0
flask~=3.1.0
pandas~=1.3.5
python-dotenv~=1.1.0
flask_sqlalchemy==3.0.3
numpy==1.21.4
lxml==4.9.3
psycopg2-binary==2.9.9
gunicorn